TaskLength interface and TaskManagerImpl implementation actualized (we've never used long type, and TimeUnti is now DateFrameable)

+++ ganttproject/src/net/sourceforge/ganttproject/task/TaskLength.java
@@ -10,7 +10,7 @@ import net.sourceforge.ganttproject.time.TimeUnit;
 public interface TaskLength {
     float getLength(TimeUnit unit);
 
-    long getLength();
+    int getLength();
 
     TimeUnit getTimeUnit();
